requirements.txt
# General Python dependencies
pandas
numpy
pyyaml

# Workflow Orchestration
apache-airflow

# Data Ingestion & Streaming
kafka-python  # Kafka Python client
requests

# Data Processing
pyspark
dbt-core

# Database & Storage
psycopg2-binary  # PostgreSQL connector
snowflake-connector-python
sqlalchemy

# Machine Learning (Optional)
scikit-learn
mlflow

# Monitoring & Logging
prometheus-client
loguru

# Synthetic Data Creation
faker  

# Kafka & Zookeeper (for easier local setup)
confluent-kafka  # Official Kafka client library (alternative to kafka-python)
